fin de exercice 4
