+++ classes/Component/AbstractComponent.php
@@ -4,38 +4,9 @@ namespace Component;
 
 use Interfaces\HasBrandInterface;
 use Interfaces\HasNameInterface;
+use Traits\HasNameTrait;
+use Traits\HasBrandTrait;
 
-abstract class AbstractComponent implements HasNameInterface, HasBrandInterface
+abstract class AbstractComponent implements HasNameInterface, HasBrandInterface, HasNameTrait, HasBrandTrait
 {
-
-    /**
-     * @var string
-     */
-    protected $name;
-    /**
-     * @var string
-     */
-    protected $brand;
-
-    public function getName(): ?string
-    {
-        return $this->name;
-    }
-
-    public function setName($n): HasNameInterface
-    {
-        $this->name = $n;
-        return $this;
-    }
-
-    public function getBrand(): ?string
-    {
-        return $this->brand;
-    }
-
-    public function setBrand($b): HasBrandInterface
-    {
-        $this->brand = $b;
-        return $this;
-    }
 }
